[Code Quality] Extract ProcessInspector and StatusFileReader from ServerManager (#211)
